pyproject.toml
[project]
name = "mcp-datapi"
version = "0.1.0"
description = "MCP Server for CDS API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["datapi>=0.4.0", "mcp[cli]>=1.6.0", "python-dotenv>=1.1.0"]
